requirements.txt
# dependencies used in all environments

# Non-Python dependencies
# sudo apt-get install python3, supervisor

                                   # LICENSE  Python3
django                             # MIT      y
django-redis                       # BSD      y
django-redis-sessions              # BSD      y
django-registration                # BSD      y
djangorestframework                # BSD      y
gunicorn                           # MIT      y
redis                              # MIT      y
requests

bpython
